requirements.txt
Flask==2.0.2
firebase-admin==5.2.0
face-recognition==1.3.0
requests==2.26.0
Pillow==8.4.0
numpy==1.21.0
